% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exbytx_hg19}
\alias{exbytx_hg19}
\title{An GRangesList created by exonsBy(txdb).}
\format{A GRangesList object with length 54070.}
\usage{
exbytx_hg19
}
\description{
The txdb is imported from the gtf file downloaded from the hg19 refSeq archive-2015-07-17-14-32-32.
}
\keyword{datasets}
