% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exbytx_mm10}
\alias{exbytx_mm10}
\title{An GRangesList created by exonsBy(txdb).}
\format{A GRangesList object with length 35119.}
\usage{
exbytx_mm10
}
\description{
The txdb is imported from the gtf file downloaded from the mm10 refSeq archive-2015-07-17-14-33-26.
}
\keyword{datasets}
