% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_seq_extraction.R
\name{fulllength}
\alias{fulllength}
\alias{tx_seq_extraction}
\alias{newf}
\title{Extract the full length transcripts sequences}
\usage{
tx_seq_extraction(BS_genome, TXDB, Write_PATH = ".",
  small_threashold = 4500, middle_threashold = 8000, trim_wsize = 2000,
  trim_ssize = 1000, divide_num = 30)

newf(x)
}
\arguments{
\item{BS_genome}{the \code{BSgenome} object of your organism.}

\item{TXDB}{the \code{txdb} object of your organism.}

\item{Write_PATH}{the path to write the transcript sequences.}

\item{small_threashold}{the transcript length threshold to put into the small group.}

\item{middle_threashold}{the transcript length threshold to put into the middle group.}

\item{trim_wsize}{the window size for the trimming.

By default, only the transcripts in the large group are trimmed.}

\item{trim_ssize}{the step size for the trimming.}

\item{divide_num}{the number of fasta files divided for each of the 3 groups. 

This argument is useful when conducting the parallel computation of the \code{RNAfold} step.}
}
\value{
This function will create 3 folders named Small, Middle, and Big. 
Each of them will contains a number of fasta files containing the full length transcripts of the sizes of each group. 

The number of the fasta files is determined by the argument \code{divide_num}.
}
\description{
\code{tx_seq_extraction} is used to extract the fasta files of the full length transcripts.
}
\examples{
\dontrun{
library(BSgenome.Mmusculus.UCSC.mm10)
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
txdb <- TxDb.Mmusculus.UCSC.mm10.knownGene
tx_seq_extraction(MMusculus,txdb,getwd())
}
}
\seealso{
\code{\link{RNAfold}}, \code{\link{Single_RNAfold}}, and \code{\link{rfold_assembly_tx}}
}
