% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBS_2ndStructure_Filter.R
\name{rBS_2ndStructure_Filter}
\alias{rBS_2ndStructure_Filter}
\title{RNA secondary structures filtering}
\usage{
rBS_2ndStructure_Filter(rBS_gr, Genome, RNA_struc_dir = NULL, TXDB = NULL)
}
\arguments{
\item{rBS_gr}{A \code{GRanges} object containing the methylation site information of RNA bisulfite sequencing.

 If the GRanges has a logical vector in its first column of the meta data column indicating the methylation states, 
 a report will be produced based on the association between the methylation and the RNA secondary structure.}

\item{Genome}{A character string which indicates the genome assembly used in rBS_gr, it can be either "mm10" or "hg19".}

\item{RNA_struc_dir}{Optinal; the directory of the self generated rds file containing the RNA secondary structures.}

\item{TXDB}{Optinal; the txdb object provided by the users.}
}
\value{
A filtered GRanges object that has no overlapping with the predicted RNA structures.
The output ranges are also filtered by the overlapping with exons or the mitochondria chromosome.
}
\description{
\code{rBS_2ndStructure} is used to filter thermal stable RNA secondary structures from RNA bisulfite sequencing data.
}
\details{
The RNA secondary structures are predicted by RNAfold in ViennaRNA package. 
The structures are MEA secondary structures predicted under gammar = 0.1 and temperature = 70 degree.

The maximum pairing distance is set to be 150 nt. At the current version of the function, 
only structures on exons and mitochondria chromosome are included. Therefore, the sites not overlapped with exons or chrM will also be filtered.
}
\examples{
rBS_gr = GRanges(seqnames = rBS_df$`#SeqID`, strand = rBS_df$refStrand, ranges = IRanges(start = rBS_df$refPos, width = 1))
rBS_gr$mcols = p.adjust(rBS_df$`p-value_mState`,method = "BH") < .05
rBS_gr_filtered <- rBS_2ndStructure_Filter(rBS_gr,"hg19")

}
\seealso{
To directly use the \code{GRangesList} of RNA secondary structures: \code{Struc_mm10},  \code{Struc_hg19}.

To generate RNA secondary structures with self specified genome sequence and transcript annotations:
 \code{\link{tx_seq_extraction}},  \code{\link{RNAfold}}, and  \code{\link{rfold_assembly_tx}}.
}
