% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rBS_df}
\alias{rBS_df}
\title{An example table of RNA bisulfite sequencing data processed by MeRanTK.}
\format{A data.frame object with 100000 rows and 20 columns:
\describe{
  #SeqID : the chromosomal IDs provided by the reference trasncripts of hg19.
  refStrand : the strands of the sites.
  refPos: the location of the sites.
  p-value_mState: the hypergeometric p value related to the statistical significance of the methylation state.
 ...
 }}
\source{
\url{http://icbi.at/software/meRanTK/}
}
\usage{
rBS_df
}
\description{
The data comes from a subset of MeRanTK processed RNA bisulfite sample of human MCF10A cell.
The table is directly imported from the txt file generated by the MeRanCall of MeRanTK with minimum read coverage of 10 and maximum conversion rate of 1.
}
\keyword{datasets}
