% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAfold.R
\name{RNAfold}
\alias{RNAfold}
\title{Parallel computation of RNAfold}
\usage{
RNAfold(Group_nums = 1:30, size = "Small", OutPutNames = "Foo")
}
\arguments{
\item{Group_nums}{This value corresponds to the \code{divide_num} argument of the \code{\link{tx_seq_extraction}}.}

\item{size}{The group name for the transcripts. Can be one of the "Small", "Middle", and "Large".}

\item{OutPutNames}{The output names of the RNAfold.}
}
\description{
\code{RNAfold} is used to submit the RNAfold command in ViennaRNA package into multiple bach jobs.
 This function is only useful when you have the qsub command and the ViennaRNA package installed on your linux system. 
 
 Other wise, please run RNAfold individually on your terminal based on the fasta files extracted by 
 \code{\link{tx_seq_extraction}}. 
 You can use \code{\link{Single_RNAfold}} to get the code of your RNAfold command.
}
\examples{
\dontrun{ 
tx_seq_extraction(MMusculus,txdb,getwd()) 
RNAfold(1:30,"Small","gsmall") 
}
}
\seealso{
\code{\link{tx_sea_extraction}}, \code{\link{Single_RNAfold}}, and \code{\link{rfold_assembly}}
}
