% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_RNAfold.R
\name{Single_RNAfold}
\alias{Single_RNAfold}
\title{Generate the bash command of RNAfold}
\usage{
Single_RNAfold(Group_num, Size = "Small", outPutNames = "Foo")
}
\arguments{
\item{Group_num}{The number of the group of the fasta file.
This value corresponds to the \code{divide_num} argument of the \code{\link{tx_seq_extraction}}.}

\item{Size}{The size of the transcript. 
Can be one of the "Small", "Middle", and "Large".}

\item{outPutNames}{The name of the output files of RNA fold.}
}
\value{
This function will return the code of RNAfold that can directly run in the terminal.
}
\description{
RNAfold is a RNA secondary structure prediction software in ViennaRNA package. 
This R function is used to generate the single bash command of RNAfold, which is used to predict the MEA RNA secondary structures of the full length transcripts.
}
\examples{
cat(Single_RNAfold(1,"Small","gsmall"))

\dontrun{
tx_seq_extraction(MMusculus,txdb,getwd()) 
system(Single_RNAfold(1,"Small","gsmall"))
}

}
\seealso{
\code{\link{tx_seq_extraction}}, \code{\link{RNAfold}}, and \code{\link{rfold_assembly}}
}
